requirements.txt
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
cssselect2==0.7.0
deprecation==2.1.0
Flask==3.0.3
fonttools==4.53.1
functions-framework==3.8.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyphen==0.15.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tinycss2==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
zopfli==0.2.3
